requirements.txt
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
gensim==3.8.3
numpy==1.18.5
wordcloud==1.8.0
textblob==0.15.3
nltk==3.5
streamlit==0.69.2
rake_nltk==1.0.4
spacy==2.3.2
spacy_streamlit==0.1.0
seaborn==0.11.0
joblib==0.17.0
sumy==0.8.1
pandas==1.1.3
matplotlib==3.3.2
Pillow==8.3.2
scikit_learn==0.24.1